requirements.txt
duckdb==1.3.2
fastapi==0.115.12
gensim==4.3.3
lightgbm==4.6.0
networkx==3.4.2
numpy==1.26.4
pandas==2.2.3
pyarrow==19.0.1
scikit-learn==1.6.1
tqdm==4.67.1
umap-learn==0.5.7
uvicorn[standard]==0.34.0
xgboost==2.1.4